requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aioping==0.4.0
aiosignal==1.4.0
anvil-app-server==1.14.0
anvil-uplink==0.6.0
arrow==1.4.0
async-timeout==5.0.1
asyncio-socks-server==0.3.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
click==8.1.8
Flask==3.1.2
frozenlist==1.8.0
future==1.0.0
greenlet==3.2.4
importlib_metadata==8.7.0
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpointer==2.1
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.0.2
pandas==2.3.3
paramiko==4.0.0
peewee==3.18.2
pillow==11.3.0
progressbar2==4.5.0
propcache==0.4.1
psutil==7.1.1
psycopg2-binary==2.9.11
pycares==4.11.0
pychrome==0.2.4
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
qrcode==8.2
six==1.17.0
SQLAlchemy==2.0.44
sqlite-web==0.6.5
tzdata==2025.2
websocket-client==1.9.0
Werkzeug==3.1.3
ws4py-sslupdate==0.5.1b0
yarl==1.22.0
yt-dlp==2025.10.14
zipp==3.23.0
